+++ lib/ex_debug_toolbar.ex
@@ -14,8 +14,8 @@ defmodule ExDebugToolbar do
     children = [
       # Start the endpoint when the application starts
       supervisor(ExDebugToolbar.Endpoint, []),
+      supervisor(ExDebugToolbar.Database.Supervisor, []),
       # Start your own worker by calling: ExDebugToolbar.Worker.start_link(arg1, arg2, arg3)
-      worker(ExDebugToolbar.Database.RequestRepo, []),
     ]
 
     # See http://elixir-lang.org/docs/stable/elixir/Supervisor.html
