+++ lib/ex_debug_toolbar/database.ex
@@ -1,8 +1,11 @@
 use Amnesia
 
 defdatabase ExDebugToolbar.Database do
+  use GenServer
+
   alias ExDebugToolbar.Data.Timeline
   alias ExDebugToolbar.Database.Request
+  alias ExDebugToolbar.Database
 
   deftable Request,
     [
@@ -17,4 +20,19 @@ defdatabase ExDebugToolbar.Database do
     type: :set,
     copying: :memory,
     index: [:id] do end
+
+  def start_link do
+    GenServer.start_link(__MODULE__, [], name: __MODULE__)
+  end
+
+  def init(_) do
+    Process.flag(:trap_exit, true)
+    [:ok, :ok] = Database.create
+    {:ok, nil}
+  end
+
+  def terminate(reason, _state) do
+    Database.destroy
+    reason
+  end
 end
